make changes to accomodate for test_add_last_time_indexes, and test_create_entity_with_make_index. Move koalas related utils instead of testing utils
