+++ featuretools/tests/testing_utils/__init__.py
@@ -2,4 +2,3 @@
 from .mock_ds import make_ecommerce_entityset
 from .features import feature_with_name, backward_path, forward_path
 from .cluster import MockClient, mock_cluster, get_mock_client_cluster
-from .misc import pd_to_ks_clean
